+++ packages/apps/manufacturing/src/components/organisms/index.js
@@ -19,17 +19,18 @@
 export {default as ConsumedProductCard} from './ConsumedProductCard/ConsumedProductCard';
 export {default as ConsumedProductGlobalCard} from './ConsumedProductGlobalCard/ConsumedProductGlobalCard';
 export {default as ManufacturingOrderCard} from './ManufacturingOrderCard/ManufacturingOrderCard';
+export {default as ManufacturingOrderHalfLabelCardList} from './ManufacturingOrderHalfLabelCardList/ManufacturingOrderHalfLabelCardList';
 export {default as ManufacturingOrderHeader} from './ManufacturingOrderHeader/ManufacturingOrderHeader';
+export {default as ManufacturingOrderIconButtonList} from './ManufacturingOrderIconButtonList/ManufacturingOrderIconButtonList';
+export {default as ManufacturingOrderNotesCardList} from './ManufacturingOrderNotesCardList/ManufacturingOrderNotesCardList';
+export {default as ManufacturingOrderProductionOrderSetView} from './ManufacturingOrderProductionOrderSetView/ManufacturingOrderProductionOrderSetView';
+export {default as ManufacturingOrderSaleOrderSetView} from './ManufacturingOrderSaleOrderSetView/ManufacturingOrderSaleOrderSetView';
 export {default as OperationOrderCard} from './OperationOrderCard/OperationOrderCard';
 export {default as OperationOrderDetailsCard} from './OperationOrderDetailsCard/OperationOrderDetailsCard';
 export {default as OperationOrderHeader} from './OperationOrderHeader/OperationOrderHeader';
+export {default as OperationOrderLabelTextList} from './OperationOrderLabelTextList/OperationOrderLabelTextList';
 export {default as ProducedProductCard} from './ProducedProductCard/ProducedProductCard';
 export {default as ProductionFileLargeCard} from './ProductionFileLargeCard/ProductionFileLargeCard';
 export {default as ProductionFileSmallCard} from './ProductionFileSmallCard/ProductionFileSmallCard';
 export {default as WasteProductCard} from './WasteProductCard/WasteProductCard';
-export {default as ManufacturingOrderIconButtonList} from './ManufacturingOrderIconButtonList/ManufacturingOrderIconButtonList';
-export {default as ManufacturingOrderOrderSetList} from './ManufacturingOrderOrderSetList/ManufacturingOrderOrderSetList';
-export {default as ManufacturingOrderHalfLabelCardList} from './ManufacturingOrderHalfLabelCardList/ManufacturingOrderHalfLabelCardList';
-export {default as ManufacturingOrderNotesCardList} from './ManufacturingOrderNotesCardList/ManufacturingOrderNotesCardList';
-export {default as OperationOrderLabelTextList} from './OperationOrderLabelTextList/OperationOrderLabelTextList';
 export {default as WasteProductFixedItems} from './WasteProductFixedItems/WasteProductFixedItems';
