+++ packages/core/src/i18n/translations/en.json
@@ -23,6 +23,9 @@
   "Base_Add": "Add",
   "Base_Check": "Check",
   "Base_Start": "Start",
+  "Base_Pause": "Pause",
+  "Base_Finish": "Finish",
+  "Base_Continue": "Continue",
   "Base_Complete": "Complete",
   "Base_Confirm": "Confirm",
   "Base_Cancel": "Cancel",
