+++ packages/core/src/i18n/translations/fr.json
@@ -23,6 +23,9 @@
   "Base_Add": "Ajouter",
   "Base_Check": "Vérifier",
   "Base_Start": "Commencer",
+  "Base_Pause": "Mettre en pause",
+  "Base_Finish": "Terminer",
+  "Base_Continue": "Continuer",
   "Base_Complete": "Terminer",
   "Base_Confirm": "Confirmer",
   "Base_Cancel": "Annuler",
